+++ include/App/App.hpp
@@ -1,7 +1,10 @@
 #ifndef RDT_APP_HPP_
 #define RDT_APP_HPP_
 
-#include "Server/ServerProtocolHandler.hpp"
-#include "Client/ClientProtocolHandler.hpp"
+#include "Tools/InterfacePerformance.hpp"
+#include "Tools/Colors.hpp"
+#include "Tools/Fixer.hpp"
+#include "Server/ServerInterface.hpp"
+#include "Client/ClientInterface.hpp"
 
 #endif//RDT_APP_HPP_
