--- include/App/Server/ServerProtocolHandler.hpp
@@ -1,15 +1,0 @@
-#ifndef SERVER_PROTOCOL_HANDLER_HPP_
-#define SERVER_PROTOCOL_HANDLER_HPP_
-
-#include <thread>
-
-namespace app
-{
-
-  class MainServerApp {
-    
-  };
-
-}
-
-#endif//SERVER_PROTOCOL_HANDLER_HPP_
