+++ include/App/Tools/Colors.hpp
@@ -13,9 +13,7 @@
 namespace tool
 {
 
-  void ConsolePrint(const char* message, const char* STYLE_CONFIG) {
-    std::cout << STYLE_CONFIG << message << NO_COLOR << std::endl;
-  }
+  void ConsolePrint(const char* message, const char* STYLE_CONFIG);
 
 }
