--- include/App/TransportParser/MainServer0SlaveParser.hpp
@@ -1,13 +1,0 @@
-#ifndef MAIN_SERVER_0_SLAVE_PARSER_HPP_
-#define MAIN_SERVER_0_SLAVE_PARSER_HPP_
-
-/*Implementation(if required)*/
-
-namespace trlt
-{
-
-
-
-}
-
-#endif//MAIN_SERVER_0_SLAVE_PARSER_HPP_
