+++ src/App/Server/ServerInterface.cpp
@@ -1,0 +1,3 @@
+#include "App/Server/ServerInterface.hpp"
+
+using namespace app;
