+++ src/App/Tools/Colors.cpp
@@ -1,0 +1,7 @@
+#include "App/Tools/Colors.hpp"
+
+using namespace tool;
+
+void ConsolePrint(const char* message, const char* STYLE_CONFIG) {
+  std::cout << STYLE_CONFIG << message << NO_COLOR << std::endl;
+}
