+++ src/MainServer.cpp
@@ -1,13 +1,13 @@
-// MainServer.cpp
-#include "Network/Socket.hpp"
+#include "Network/UdpSocket.hpp"
+#include "Server/ServerInterface.hpp"
+
 int main(){
-  RDT::UdpSocket mainServerListener("", "8000");
-  RDT::UdpSocket slaveServerSocket("35.188.208.43", "8000");
+  MainServerApp MainServer; 
+  net::UdpSocket mainServerListener("", "8000");
+  net::UdpSocket slaveServerSocket("35.188.208.43", "8000");
   std::string received_message, IP_from;
   uint16_t Port_from;
-  // recibes el mensaje desde el cliente
   mainServerListener.receive(received_message, IP_from, Port_from);
-  // envías el mensaje al servidor esclavo
   slaveServerSocket.send(received_message);
   return 0;
 }
