+++ src/Network/RDTMask.cpp
@@ -1,0 +1,3 @@
+#include "Network/RDTMask.hpp"
+
+using namespace rdt;
